cluster_task_manager conflicts with GCS

+++ src/ray/core_worker/transport/direct_task_transport.h
@@ -310,7 +310,7 @@ class CoreWorkerDirectTaskSubmitter {
     // Check whether it's safe to delete this SchedulingKeyEntry from the
     // scheduling_key_entries_ hashmap.
     inline bool CanDelete() const {
-      if (pending_lease_requests.empty() && task_priority_queue.empty() &&
+      if ((long unsigned int)pending_lease_requests.empty() && task_priority_queue.empty() &&
           active_workers.size() == 0 && num_busy_workers == 0) {
         return true;
       }
@@ -326,7 +326,7 @@ class CoreWorkerDirectTaskSubmitter {
 
     // Get the current backlog size for this scheduling key
     [[nodiscard]] inline int64_t BacklogSize() const {
-      if (task_priority_queue.size() < pending_lease_requests.size()) {
+      if ((long unsigned int)task_priority_queue.size() < pending_lease_requests.size()) {
         // This can happen if worker is reused.
         return 0;
       }
